moving the FibonacciBinder.kt file from the binder package to the fibonacci package
